refactor(arfs cache): Rename file PE-753
